requirements.txt
alabaster==0.7.12
asgiref==3.3.0
Babel==2.9.0
blis==0.4.1
boto3==1.4.7
botocore==1.7.43
catalogue==2.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudmersive-ocr-api-client==3.0.1
colorama==0.4.4
cycler==0.10.0
cymem==2.0.4
Cython==0.29.14
defusedxml==0.5.0
dill==0.3.3
Django==3.0.8
django-multiforloop==0.2.1
django-storages==1.10.1
djangorestframework==3.11.0
docutils==0.14
EMD-signal==0.2.13
en-core-web-md==2.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
gensim==3.8.3
html5lib==0.999999999
httplib2==0.18.1
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.3
joblib==0.17.0
jws==0.1.3
kiwisolver==1.3.1
language-tool-python==2.4.6
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
multiprocess==0.70.11.1
murmurhash==1.0.4
mysql-connector==2.2.9
mysqlclient==1.4.6
nltk==3.5
numpy==1.20.2
numpydoc==1.1.0
oauth2client==3.0.0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.4
pathos==0.2.7
pathy==0.3.5
pbr==4.0.2
Pillow==7.2.0
plac==1.1.3
pox==0.2.9
ppft==1.6.6.3
preshed==3.0.4
psycopg2==2.7.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
pyemd==0.5.1
Pygments==2.8.1
pymed==0.8.9
pymongo==3.10.1
PyMuPDF==1.18.6
pyparsing==2.4.7
pysbd==0.3.3
pytesseract==0.3.7
python-dateutil==2.7.3
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2017.3
regex==2020.11.11
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.6
s3transfer==0.1.11
scikit-learn==0.24.1
scipy==1.5.4
six==1.10.0
smart-open==3.0.0
snowballstemmer==2.1.0
spacy==3.0.1
spacy-legacy==3.0.1
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
srsly==2.4.0
stevedore==1.28.0
thinc==8.0.1
threadpoolctl==2.1.0
tqdm==4.51.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.25.9
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wasabi==0.8.2
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.4.0
